extract serverservice client code for general use
